fix(server): add missing fields and update types (#10)
* Changed variable name for problems sql file

* Change some variables and sync with table inputs

* Change variable data type according to query requirements

* Change times to YEAR

* Remove sqlnet.log

* Change time in datatype and added labour hours to Repair jobs

closes #7, closes #8

+++ web/migrations/002_create_parts.sql
@@ -1,4 +1,4 @@
 CREATE TABLE parts(
-	name VARCHAR(40) PRIMARY KEY,
-	cost INT         NOT NULL	
+	name VARCHAR(40)   PRIMARY KEY,
+	cost NUMBER(10, 2) NOT NULL	
 )
